Clean up case-by-index-handler test.
